+++ zan-extension/src/os/AsyncIO.c
@@ -97,7 +97,7 @@ void swAio_callback_test(swAio_event *aio_event)
 {
     printf("content=%s\n", (char *)aio_event->buf);
     printf("fd: %d, request_type: %s, offset: %lld, length: %llu\n", aio_event->fd,
-            (aio_event == SW_AIO_READ) ? "READ" : "WRITE", aio_event->offset, (uint64_t) aio_event->nbytes);
+            (aio_event == SW_AIO_READ) ? "READ" : "WRITE", (long long int)aio_event->offset, (long long unsigned) aio_event->nbytes);
     SwooleG.running = 0;
 }
